TANGERINE-197: changed Percentage to BigDecimal
git-svn-id: svn://svn.orangeleap.com/tangerine/mpower/trunk@3920 0f581344-5646-3a48-8ded-616bfea576c6

+++ src/com/mpower/domain/DistributionLine.java
@@ -44,7 +44,7 @@ public class DistributionLine implements Customizable, Viewable, Serializable {
     private BigDecimal amount;
     
     @Column(name = "PERCENTAGE")
-    private Float percentage;
+    private BigDecimal percentage;
 
     @Column(name = "PROJECT_CODE")
     private String projectCode;
@@ -116,11 +116,11 @@ public class DistributionLine implements Customizable, Viewable, Serializable {
         this.amount = amount;
     }
 
-    public Float getPercentage() {
+    public BigDecimal getPercentage() {
         return percentage;
     }
 
-    public void setPercentage(Float percentage) {
+    public void setPercentage(BigDecimal percentage) {
         this.percentage = percentage;
     }
